fixed erroe on login and sign up page by making first letter of component upper case
