+++ src/app/login/page.js
@@ -4,7 +4,7 @@ import { useSignInWithEmailAndPassword } from "react-firebase-hooks/auth"
 import { auth } from "../firebase/config";
 import { useRouter } from "next/navigation";
 
-const page = () => {
+const Page = () => {
   // State variables for form inputs
   const [formData, setFormData] = useState({
     email: "",
@@ -107,4 +107,4 @@ const page = () => {
   );
 };
 
-export default page;
+export default Page;
